+++ src/main/java/com/formation/rencontre/entities/Photo.java
@@ -14,17 +14,17 @@ public class Photo {
 @Id
 @Column(name="idphoto", unique= true, nullable= false)
 @GeneratedValue(strategy=GenerationType.AUTO)		
-private int idphoto;
+private Long idphoto;
 @Column(name="clien",nullable=false,length=10)
 private String lien;
 @Column(name="note",nullable=false,length=2)
-private int note;
+private Integer note;
 @ManyToOne
 private Utilisateur utilisateur;
-public int getIdphoto() {
+public Long getIdphoto() {
 	return idphoto;
 }
-public void setIdphoto(int idphoto) {
+public void setIdphoto(Long idphoto) {
 	this.idphoto = idphoto;
 }
 public String getLien() {
@@ -33,13 +33,13 @@ public String getLien() {
 public void setLien(String lien) {
 	this.lien = lien;
 }
-public int getNote() {
+public Integer getNote() {
 	return note;
 }
-public void setNote(int note) {
+public void setNote(Integer note) {
 	this.note = note;
 }
-public Photo(int idphoto, String lien, int note) {
+public Photo(Long idphoto, String lien, Integer note) {
 	super();
 	this.idphoto = idphoto;
 	this.lien = lien;
